% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_hyper_midP.R
\name{central_hyper_midP}
\alias{central_hyper_midP}
\title{Central hypergeometric distribution mid-P function}
\usage{
central_hyper_midP(U0, S0, S1, U1, d0)
}
\arguments{
\item{U0}{Integer, number of undetected extant species
at the previous timestep.}

\item{S0}{Integer, number of detected extant species
at the previous timestep.}

\item{S1}{Integer, number of detected extant species
at the current timestep.}

\item{U1}{Integer, number of undetected extant species
at the current timestep.}

\item{d0}{Integer, number of species detected during
the timestep.}
}
\value{
alpha: Float [0,1], analogous to a confidence level.
}
\description{
Calculates the mid-P function's value for the 
 central hypergeometric SEUX model.
}
\details{
The mid-P function is 

 \eqn{ P_M(U_b \geq U^*|\phi) = 0.5 [ P(\Phi \leq \phi | U^*) + 1 - P(\Phi \geq \phi | U^*) ] }.

 Sampling according the mid-P function is equivalent to sampling fuzzy confidence bounds.
}
\examples{
 S0 <- 100; S1 <- 96; d0 <- 1    # from the data
 U1 <- 200;                      # from an earlier timestep in the algorithm
 central_hyper_midP(211, S0, S1, U1, d0) # 0.5491703
 central_hyper_midP(212, S0, S1, U1, d0) # 0.4838616

}
