% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frst_last_detns.R
\name{check_frst_last_detns}
\alias{check_frst_last_detns}
\title{Check first and last detections}
\usage{
check_frst_last_detns(frstDetn, lastDetn)
}
\arguments{
\item{frstDetn}{Vector of integers, species' year of first detection.}

\item{lastDetn}{Vector of integers, species' year of last detection.}
}
\value{
A list with elements: flag, message.
 \describe{
 \item{flag:}{Boolean, is TRUE if the data passed the checks.}
 \item{message:}{String, a message explaining why the data didn't pass the checks (can be
  used e.g. in the stop function).}
 }
}
\description{
Performs various checks on the first and last detection dates
 -- e.g. first detection is always less than or equal to last detection --
 to ensure that they make sense.
}
