% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_last_detections_from_csv.R
\name{get_first_last_detections_from_csv}
\alias{get_first_last_detections_from_csv}
\title{Get first and last detections from csv file}
\usage{
get_first_last_detections_from_csv(fName, frst_col, last_col)
}
\arguments{
\item{fName}{String, a filename.}

\item{frst_col}{Integer or String, the column containing each species' year of first detection.}

\item{last_col}{Integer or String, the column containing each species' year of last detection.}
}
\value{
A dataframe with two columns: frstDetn, lastDetn
 \describe{
  \item{frstDetn:}{Integers, species' year of first detection.}
  \item{lastDetn:}{Integers, species' year of last detection.}
 }
}
\description{
Returns the first and last detection years
 stored in a csv file after performing basic checks.
}
