% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_inputs.R
\name{get_model_inputs}
\alias{get_model_inputs}
\title{Get SEUX model inputs}
\usage{
get_model_inputs(frstDetn, lastDetn, collapse_timesteps = TRUE,
  y0 = NULL, yf = NULL)
}
\arguments{
\item{frstDetn}{Vector of integers, species' year of first detection.}

\item{lastDetn}{Vector of integers, species' year of last detection.}

\item{collapse_timesteps}{Boolean. If set to FALSE, each year is treated as a timestep
in the model. If set to TRUE, timesteps are collapsed together so that each timestep
has at least one detected extinction. Defaults to TRUE.}

\item{y0}{Integer, the start year for the timeseries.
Default is the year of the first detected extinction.}

\item{yf}{Integer, the final year for the timeseries.
Default is the year of the last new discovery or the year of the last extinction,
whichever is later.}
}
\value{
A dataframe with four columns: year, S, E, d
 \describe{
  \item{year:}{Integers, the year to which each timestep corresponds.}
  \item{S:}{Integers, number of detected extant species at each timestep.}
  \item{E:}{Integers, number of detected extinct species at each timestep.}
  \item{d:}{Integers, number of species discovered at each timestep.}
 }
}
\description{
Turns first and last detection dates into input for the SEUX model.
}
\details{
This function is mainly to provide a convenient way to impose constraints on the
 timeseries extent and collapse timesteps.
}
