% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_old_estimate.R
\name{get_old_estimate}
\alias{get_old_estimate}
\title{Get old SEUX estimates}
\usage{
get_old_estimate(S, E, U_T = 0)
}
\arguments{
\item{S}{Vector of integers, number of detected extant species
at each timestep.}

\item{E}{Vector of integers, number of detected extinct species
at each timestep.}

\item{U_T}{Integer, the number of undetected extant species remaining at the end of the
timeseries. Typically an assumption. Default value is 0.}
}
\value{
A dataframe with three columns: U_old, X_old, mu_estimate
 \describe{
  \item{U_old:}{Floats, an estimate of the number of undetected extant species at each timestep.}
  \item{X_old:}{Floats, an estimate of the number of undetected extinct species at each timestep.}
  \item{mu_estimate:}{Floats, an estimate of the extinction rate at each timestep (calculated from detected species' extinctions. Note that, at the final timestep, an estimate cannot be made (NA value).}
 }
}
\description{
Obtains the estimates of U and X from the method in Chisholm et al. (2016).
}
\details{
This function uses a simplified method for obtaining the estimates.
 The number of species discovered at the previous timestep

  \eqn{ d_{t-1} = \nu_t U_{t-1} }

 can be obtained from the data. Then the 3rd line from Eq. 1 in 
 Chisholm et al. (2016) can be rerranged

  \eqn{ U_{t-1} = ( U_t + d_{t-1} ) / ( 1 - \mu_{t-1} ) }

 Given an assumed number of undetected species remaining at the
 end of the timeseries, \eqn{U_T} (typically assumed \eqn{U_T = 0}),
 then U_t can be solved by working backwards through the timeseries.
}
